requirements.txt
langchain
fastapi
sentence-transformers
pypdf
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
streamlit
deepeval

-e .